fix : prevent keeper from not going inside penalty lol
